update deno part 2
