+++ deno_webgpu/surface.rs
@@ -2,10 +2,7 @@
 
 use super::WebGpuResult;
 use deno_core::error::AnyError;
-use deno_core::include_js_files;
 use deno_core::op;
-use deno_core::Extension;
-use deno_core::ExtensionBuilder;
 use deno_core::OpState;
 use deno_core::Resource;
 use deno_core::ResourceId;
@@ -14,40 +11,20 @@ use std::borrow::Cow;
 use std::rc::Rc;
 use wgpu_types::SurfaceStatus;
 
-fn ext() -> ExtensionBuilder {
-    Extension::builder_with_deps(
-        "deno_webgpu_surface",
-        &["deno_webidl", "deno_web", "deno_webgpu"],
-    )
-}
-
-fn ops(ext: &mut ExtensionBuilder, unstable: bool) -> &mut ExtensionBuilder {
-    ext.ops(vec![
-        op_webgpu_surface_configure::decl(),
-        op_webgpu_surface_get_current_texture::decl(),
-        op_webgpu_surface_present::decl(),
-    ])
-    .state(move |state| {
-        // TODO: check & possibly streamline this
-        // Unstable might be able to be OpMiddleware
-        // let unstable_checker = state.borrow::<super::UnstableChecker>();
-        // let unstable = unstable_checker.unstable;
-        state.put(super::Unstable(unstable));
-    })
-}
-
-pub fn init_ops_and_esm(unstable: bool) -> Extension {
-    ops(&mut ext(), unstable)
-        .esm(include_js_files!(
-            "03_surface.js",
-            "04_surface_idl_types.js",
-        ))
-        .build()
-}
-
-pub fn init_ops(unstable: bool) -> Extension {
-    ops(&mut ext(), unstable).build()
-}
+deno_core::extension!(
+    deno_webgpu_surface,
+    deps = [deno_webidl, deno_web, deno_webgpu],
+    ops = [
+        op_webgpu_surface_configure,
+        op_webgpu_surface_get_current_texture,
+        op_webgpu_surface_present,
+    ],
+    esm = ["02_surface.js"],
+    options = { unstable: bool },
+    state = |state, options| {
+        state.put(super::Unstable(options.unstable));
+    },
+);
 
 pub struct WebGpuSurface(pub crate::Instance, pub wgpu_core::id::SurfaceId);
 impl Resource for WebGpuSurface {
